Added filtering bookmarks, wrapping & UI highlighting, improved readability
Filtering now works on pages with bookmarks and correctly handles work tags and bookmarker tags. There was some additional refactoring to handle differentiation between the two, as bookmarker tags are still classed as tags (albeit meta ones).

Because of the way Ao3 handles text wrapping for tag elements, it was possible for tags to wrap onto multiple lines, making it difficult to see what you were filtering. Previous versions solved this by wrapping the tag and UI elements in a span that forced it all to render on a single line, but there were cases with longer tags where this would break formatting.

This revision uses inline scripting to change the wrapper's appearance on mouseover, adding a visual highlight so the user can easily determine what's going to be filtered if they click on something; support is provided for altering the background color, as well as providing a border.

Did some additional refactoring and cleanup: things are broken down a little more firmly into functions, and there's been a shift to take string literals and condense them down into static classes for ease of modification.